requirements.txt
numpy >= 1.11.1
Pillow >= 3.2.0
scipy >= 0.18.0